chore: Move RadioGroup component to UI directory
