+++ packages/webapp-libs/webapp-core/src/components/forms/index.ts
@@ -2,4 +2,3 @@ export * from './input';
 export * from './dropzone';
 export * from './checkbox';
 export * from './form';
-export * from './select';
