--- packages/webapp-libs/webapp-core/src/components/forms/radioGroup/index.ts
@@ -1,2 +1,0 @@
-export * from './radioGroup.component';
-export * from './radioGroupItem';
