--- packages/webapp-libs/webapp-core/src/components/forms/radioGroup/radioGroup.component.tsx
@@ -1,14 +1,0 @@
-import * as RadioGroupPrimitive from '@radix-ui/react-radio-group';
-import * as React from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const RadioGroup = React.forwardRef<
-  React.ElementRef<typeof RadioGroupPrimitive.Root>,
-  React.ComponentPropsWithoutRef<typeof RadioGroupPrimitive.Root>
->(({ className, ...props }, ref) => {
-  return <RadioGroupPrimitive.Root className={cn('grid gap-2', className)} {...props} ref={ref} />;
-});
-RadioGroup.displayName = RadioGroupPrimitive.Root.displayName;
-
-export { RadioGroup };
