--- packages/webapp-libs/webapp-core/src/components/forms/radioGroup/radioGroupItem/index.ts
@@ -1,1 +1,0 @@
-export * from './radioGroupItem.component';
